requirements.txt
# Core API
fastapi==0.110.3
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.4
pydantic-settings==2.3.4

# OCR
python-doctr[torch]==0.9.0
pytesseract==0.3.13
Pillow==10.3.0
pdf2image==1.17.0

# LLM & Vector Store
langchain==0.2.6
langchain-openai==0.1.14
langchain-anthropic==0.1.20
langchain-google-genai
langchain-groq
langchain-community==0.2.6
chromadb==0.5.3
tiktoken==0.7.0

# MLOps
mlflow==2.14.1
evidently==0.4.33

# Monitoring
prometheus-client==0.20.0
structlog==24.2.0

# Database
sqlalchemy==2.0.31
aiosqlite==0.20.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
jinja2==3.1.4
aiofiles==24.1.0

# Frontend
streamlit==1.36.0
